Update to Windows

+++ fibos/manage_os.py
@@ -14,7 +14,7 @@ IRESF = 1
 
 
 def occluded_surface(pdb,method = "FIBOS"):
-    create_folder()
+    create_folder(pdb)
     pdb = pdb.lower()
     name_pack = "fibos"
     path_pack = pkgutil.get_loader(name_pack).get_filename()
@@ -31,7 +31,18 @@ def occluded_surface(pdb,method = "FIBOS"):
         meth = 2
     else:
         print("Wrong Method")
-    meth = 1
+    file_remove = pdb
+    if pdb.endswith(".pdb"):
+        file_remove = pdb.replace(".pdb","")
+        ray_remove = "raydist_"+file_remove+".lst"
+        pack_remove = "prot_"+file_remove+".pak"
+        file_remove = "prot_"+file_remove+".srf"
+    if os.path.exists(file_remove):
+        os.remove(file_remove)
+    if os.path.exists(ray_remove):
+        os.remove(ray_remove)
+    if os.path.exists(pack_remove):
+        os.remove(pack_remove)
     main_intermediary.call_main(IRESF,iresl,MAX_RES,MAX_AT,meth)
     remove_files()
     file_name = rename_file(pdb)
@@ -49,9 +60,12 @@ def remove_files():
     files = [file for file in os.listdir(path) if any(file.endswith(ext) for ext in extensions)]
     for file in files:
         os.remove(os.path.join(path, file))
-    os.remove(os.path.join(path,"fort.6"))
-    os.remove(os.path.join(path, "part_i.pdb"))
-    os.remove(os.path.join(path, "part_v.pdb"))
-    os.remove("radii")
-    os.remove("temp.pdb")
-    os.remove("temp.cln")
+    if os.path.exists(os.path.join(path,"fort.6")):
+        os.remove(os.path.join(path,"fort.6"))
+    if os.path.exists(os.path.join(path, "part_i.pdb")):
+        os.remove(os.path.join(path, "part_i.pdb"))
+    if os.path.exists(os.path.join(path, "part_v.pdb")):
+        os.remove(os.path.join(path, "part_v.pdb"))
+    #os.remove("temp.pdb")
+    if os.path.exists("temp.cln"):
+        os.remove("temp.cln")
